creating an additional base class for RazorPage, which only contains code that's not AspNetCore / HttpContext specific
