+++ src/Mvc/Mvc.Razor/src/RazorPage.cs
@@ -5,6 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Razor;
 using Microsoft.AspNetCore.Html;
 using Microsoft.AspNetCore.Http;
 
@@ -95,7 +96,7 @@ namespace Microsoft.AspNetCore.Mvc.Razor
         /// </summary>
         /// <param name="name">The name of the section to render.</param>
         /// <returns>An empty <see cref="IHtmlContent"/>.</returns>
-        /// <remarks>The method writes to the <see cref="RazorPageBase.Output"/> and the value returned is a token
+        /// <remarks>The method writes to the <see cref="RazorRazorPageBase{TOutputContext}.Output"/> and the value returned is a token
         /// value that allows the Write (produced due to @RenderSection(..)) to succeed. However the
         /// value does not represent the rendered content.</remarks>
         public HtmlString RenderSection(string name)
@@ -114,7 +115,7 @@ namespace Microsoft.AspNetCore.Mvc.Razor
         /// <param name="name">The section to render.</param>
         /// <param name="required">Indicates if this section must be rendered.</param>
         /// <returns>An empty <see cref="IHtmlContent"/>.</returns>
-        /// <remarks>The method writes to the <see cref="RazorPageBase.Output"/> and the value returned is a token
+        /// <remarks>The method writes to the <see cref="RazorRazorPageBase{TOutputContext}.Output"/> and the value returned is a token
         /// value that allows the Write (produced due to @RenderSection(..)) to succeed. However the
         /// value does not represent the rendered content.</remarks>
         public HtmlString RenderSection(string name, bool required)
@@ -137,7 +138,7 @@ namespace Microsoft.AspNetCore.Mvc.Razor
         /// <returns>
         /// A <see cref="Task{HtmlString}"/> that on completion returns an empty <see cref="IHtmlContent"/>.
         /// </returns>
-        /// <remarks>The method writes to the <see cref="RazorPageBase.Output"/> and the value returned is a token
+        /// <remarks>The method writes to the <see cref="RazorRazorPageBase{TOutputContext}.Output"/> and the value returned is a token
         /// value that allows the Write (produced due to @RenderSection(..)) to succeed. However the
         /// value does not represent the rendered content.</remarks>
         public Task<HtmlString> RenderSectionAsync(string name)
@@ -159,7 +160,7 @@ namespace Microsoft.AspNetCore.Mvc.Razor
         /// <returns>
         /// A <see cref="Task{HtmlString}"/> that on completion returns an empty <see cref="IHtmlContent"/>.
         /// </returns>
-        /// <remarks>The method writes to the <see cref="RazorPageBase.Output"/> and the value returned is a token
+        /// <remarks>The method writes to the <see cref="RazorRazorPageBase{TOutputContext}.Output"/> and the value returned is a token
         /// value that allows the Write (produced due to @RenderSection(..)) to succeed. However the
         /// value does not represent the rendered content.</remarks>
         /// <exception cref="InvalidOperationException">if <paramref name="required"/> is <c>true</c> and the section
