+++ src/Mvc/Mvc.RazorPages/src/Microsoft.AspNetCore.Mvc.RazorPages.csproj
@@ -10,6 +10,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\..\..\Razor\Razor\src\Microsoft.AspNetCore.Razor.csproj" />
     <Reference Include="Microsoft.AspNetCore.Mvc.Razor" />
   </ItemGroup>
