+++ src/Razor/Razor/src/IRazorUrlHelper.cs
@@ -1,9 +1,11 @@
 // Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+using System.IO;
+using Microsoft.AspNetCore.Razor.TagHelpers;
 
-using System.Threading.Tasks;
-
-namespace Microsoft.AspNetCore.Mvc.Razor
+namespace Microsoft.AspNetCore.Razor
 {
-    public delegate Task RenderAsyncDelegate();
+    public interface IRazorUrlHelper
+    {
+    }
 }
