+++ src/Razor/Razor/src/RenderAsyncDelegate.cs
@@ -1,0 +1,9 @@
+// Copyright (c) .NET Foundation. All rights reserved.
+// Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
+
+using System.Threading.Tasks;
+
+namespace Microsoft.AspNetCore.Razor
+{
+    public delegate Task RenderAsyncDelegate();
+}
